+++ packages/drqp_serial/include/drqp_serial/SerialPlayer.h
@@ -18,33 +18,33 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#pragma once 
+#pragma once
 
 #include <cstddef>
-#include "drqp_serial/SerialProtocol.h"
 #include <deque>
+#include <string>
+
 #include "RecordingProxy.h"
+#include "drqp_serial/SerialProtocol.h"
 
-namespace RecordingProxy
-{
-class SerialPlayer: public SerialProtocol
-{
-public:
-    void begin(const unsigned long baudRate, const uint8_t transferConfig) override;
-    size_t write(uint8_t byte) override;
-    bool available() override;
-    uint8_t peek() override;
-    uint8_t read() override;
-    void flushRead() override;
-    size_t write(const uint8_t *data, size_t size) override;
-    size_t readBytes(uint8_t *buffer, size_t size) override;
+namespace RecordingProxy {
+class SerialPlayer : public SerialProtocol {
+ public:
+  void begin(const uint32_t baudRate, const uint8_t transferConfig) override;
+  size_t write(uint8_t byte) override;
+  bool available() override;
+  uint8_t peek() override;
+  uint8_t read() override;
+  void flushRead() override;
+  size_t write(const uint8_t* data, size_t size) override;
+  size_t readBytes(uint8_t* buffer, size_t size) override;
 
-    void load(const std::string &fileName);
-private:
-    Record currentRecord_;
-    OperationType lastOperation_;
-    std::deque<Record> records_;
-    Record &currentRecord();
+  void load(const std::string& fileName);
+
+ private:
+  Record currentRecord_;
+  OperationType lastOperation_;
+  std::deque<Record> records_;
+  Record& currentRecord();
 };
-}
-
+}  // namespace RecordingProxy
