+++ packages/drqp_serial/include/drqp_serial/TcpSerial.h
@@ -31,7 +31,7 @@ public:
   TcpSerial(const std::string &ip, uint16_t port);
 
   using SerialProtocol::begin;
-  void begin(const unsigned long baudRate,
+  void begin(const uint32_t baudRate,
              const uint8_t transferConfig) override;
 
   size_t write(uint8_t byte) override;
