+++ packages/drqp_serial/include/drqp_serial/UnixSerial.h
@@ -30,7 +30,7 @@ public:
     ~UnixSerial();
 
     using SerialProtocol::begin;
-    void begin(const unsigned long baudRate, const uint8_t transferConfig) override;
+    void begin(const uint32_t baudRate, const uint8_t transferConfig) override;
     size_t write(uint8_t byte) override;
     bool available() override;
     void flushRead() override;
